Add files via upload
Matrimonial website which will provide platform to a lot of Bride/Groom for finding perfect match. There are different sectors like Registration,Partner,Search,etc.So the Bride/Groom can get their interest for find their partner. Bride/Groom can directly search Partner according to their required criteria.
The main objective of Matrimonial Web Application is to provide Grooms and Brides with excellent matchmaking experience by exploring the opportunities and resources to meet true potential partner. Keeping our objective in mind, we have created a world renowned online matchmaking services that will touch the souls of millions of people all over the globe. Marriage is viewed differently in India as compared to the West where it is largely a matter of individual choice. In India, marriage is viewed not so much as a union between two individuals as the beginning of an enduring relationship between two families. Weddings are usually protracted events that mark the end of lengthy negotiations between two extended families including aunts, uncles, and even cousins once step removed (Seymour, 1999). Referred to as ‘arranged marriage’, they are rarely based purely on individual preference, choice or love. Marriage symbolizes and a forms the collective nature of family and larger kinship units in which the families are embedded. In contrast, the western notion of marriage labeled as ‘love marriage’ is frowned upon by the more traditional family elders (Dion & Dion, 1996). Globalization of the economy, urbanization and the increased influence of western popular culture from books to movies and television shows, have brought about changes in the society.